refactor(api): rename GoogleOidcAuthenticationService.js file
into google-oidc-authentication-service.js
