+++ api/config/setup-oidc-authentication-service-registry.js
@@ -2,7 +2,7 @@ import { logger } from '../src/shared/infrastructure/utils/logger.js';
 import { oidcAuthenticationServiceRegistry } from '../lib/domain/services/authentication/authentication-service-registry.js';
 import { CnavOidcAuthenticationService } from '../lib/domain/services/authentication/cnav-oidc-authentication-service.js';
 import { FwbOidcAuthenticationService } from '../lib/domain/services/authentication/fwb-oidc-authentication-service.js';
-import { GoogleOidcAuthenticationService } from '../lib/domain/services/authentication/GoogleOidcAuthenticationService.js';
+import { GoogleOidcAuthenticationService } from '../lib/domain/services/authentication/google-oidc-authentication-service.js';
 import { PaysdelaloireOidcAuthenticationService } from '../lib/domain/services/authentication/paysdelaloire-oidc-authentication-service.js';
 import { PoleEmploiOidcAuthenticationService } from '../lib/domain/services/authentication/pole-emploi-oidc-authentication-service.js';
